refactored RTMAPI error and added higer level wrapper for more meaningful nested errors
