+++ Sources/RTMAPI/RTMAPIEvent+Builders.swift
@@ -12,7 +12,7 @@ import Common
 
 //MARK: - Protocol
 /// An abstraction representing an object capable of building a realtime messaging api event
-protocol RTMAPIEventBuilder {
+public protocol RTMAPIEventBuilder {
     /// The name of the event `type`s this object builds
     static var eventTypes: [String] { get }
     
@@ -50,30 +50,6 @@ extension RTMAPIEventBuilder {
     static func canMake(fromJson json: JSON) -> Bool {
         guard let event = json["type"].string else { return false }
         return (self.eventTypes.contains(event))
-    }
-}
-
-//MARK: - Errors
-/// Describes a range of errors that can occur when attempting to build models from `JSON`
-public enum RTMAPIEventBuilderError: ErrorProtocol, CustomStringConvertible {
-    /// No builder for the provided type exists
-    case noAvailableBuilder(type: String)
-    
-    /// The builder in not capable of handling the provided `JSON`
-    case invalidBuilder(builder: Any.Type)
-    
-    /// The response was invalid or the data was unexpected
-    case invalidResponse(json: JSON)
-    
-    public var description: String {
-        switch self {
-        case .invalidBuilder(let builder):
-            return "The chosen builder: '\(String(builder.self))' cannot handled to provided data"
-        case .invalidResponse(let json):
-            return "The response was invalid:\n\(json.jsonValueDescription)"
-        case .noAvailableBuilder(let type):
-            return "There is no builder available for the event-type: \(type)"
-        }
     }
 }
